### 2020-12-05 #2889 @whedon
[PRE REVIEW]: SymbiotaR2: An R Package for Accessing Symbiota2 Data
**Submitting author:** @akoontz11 (<a href="http://orcid.org/0000-0002-6103-5894">Austin Koontz</a>)
**Repository:** <a href="https://github.com/ropensci/SymbiotaR2" target ="_blank">https://github.com/ropensci/SymbiotaR2</a>
**Version:** v1.0.0
