### 2020-12-05 #2889 @kyleniemeyer
Hi @akoontz11, thanks for your interest in JOSS. The editorial board is going to review your software to see if it falls within our scope—in particular, while useful, this looks like an API client rather than [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software).

I did note that this comes from rOpenSci; if deemed to be within the scope of JOSS, the review/publication will be fast-tracked, but not all rOpenSci-approved packages are automatically within the scope of JOSS.
