### 2020-12-15 #2889 @akoontz11
Hi @kyleniemeyer, thanks for your message and providing information on this submission. This software has completed [review by rOpenSci](https://github.com/ropensci/software-review/issues/373), but I wanted to reach out and just ask about the status of this review. If there's anything I could do to help move the process along, please let me know.
