### 2020-12-16 #2917 @kyleniemeyer
rOpenSci review: https://github.com/ropensci/software-review/issues/373
